refactor: remove admin naming from attraction-related pages
